# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IELTSVega Mobile API",
    "version": "1.0.0",
    "description": "The JSON API behind the IELTSVega mobile app.\n\n**Envelope.** Every response is `{ ok: true, data }` or `{ ok: false, error }`. Branch on `ok`, and on `error.code` — not on the HTTP status, which is a coarser view of the same thing.\n\n**Auth.** `Authorization: Bearer <token>`, from any of the sign-in endpoints. Send `X-Client-Platform: ios|android` so sessions are scoped per platform — without it the app's session counts as a web one and will evict the candidate's browser session.\n\n**Scores are never sent by the client.** Answers go up, bands come back."
  },
  "servers": [
    {
      "url": "https://ieltsvega.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/dashboard": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Everything the home screen draws",
        "responses": {
          "200": {
            "description": "Stats, usage, entitlements and focus.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok",
                    "data"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/practice/recording": {
      "post": {
        "tags": [
          "practice"
        ],
        "summary": "Upload one speaking take",
        "description": "`multipart/form-data`, field `audio`. MUST be 16 kHz mono PCM WAV — what Flutter's `record` package produces natively. Returns an audio location, never a band.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "audio"
                ],
                "properties": {
                  "audio": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stored.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok",
                    "data"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A session token."
      }
    },
...
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "unauthenticated",
              "forbidden",
              "plan_required",
              "validation_failed",
              "not_found",
              "conflict",
              "rate_limited",
              "payload_too_large",
              "service_unavailable",
              "server_error"
            ]
          },
          "message": {
            "type": "string",
            "description": "Safe to show a user as-is."
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Per-field messages, keyed as the request body. Validation only."
          },
          "retryAfterSec": {
            "type": "number"
          },
          "plan": {
            "type": "object",
            "additionalProperties": true,
            "description": "On `plan_required`: which tier, why, and how much allowance is spent."
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "The failure half of the envelope.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "ok",
                "error"
              ],
              "properties": {
                "ok": {
                  "type": "boolean",
                  "enum": [
                    false
                  ]
                },
                "error": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  }
}